refact: view friends 에서 User로 이동
