+++ backend/friends/views.py
@@ -1,11 +1,9 @@
 from django.contrib.auth import get_user_model
 from django.db.models import Q
 from rest_framework.views import APIView
-from rest_framework.generics import ListAPIView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from drf_spectacular.utils import extend_schema, OpenApiResponse, OpenApiParameter
-from rest_framework import filters
 from .serializers import FriendRequestWithOtherUserSerializer
 from rest_framework.generics import GenericAPIView
 
@@ -13,7 +11,6 @@ from .models import Friend
 from .serializers import (
     FriendSerializer,
     FriendRequestSerializer,
-    UserSearchSerializer,
 )
 from .error import FriendError
 from .detail import FriendDetail
@@ -29,10 +26,8 @@ User = get_user_model()
 @extend_schema(
     tags=["Friends"],
 )
-
 class FriendRequestActionView(APIView):
     permission_classes = [IsAuthenticated]
-
 
     @extend_schema(
         summary="Accept a Friend Request",
@@ -43,7 +38,6 @@ class FriendRequestActionView(APIView):
             200: OpenApiResponse(
                 description="Friend request accepted successfully.",
                 response=FriendRequestWithOtherUserSerializer,
-
             ),
             404: OpenApiResponse(
                 description="User or friend request not found.", response=str
@@ -119,6 +113,7 @@ class FriendRequestActionView(APIView):
         # Return a success message
         return Response({"message": FriendDetail.REQUEST_REJECTED.value}, status=200)
 
+
 @extend_schema(
     summary="List Friends",
     description="List friends for the authenticated user.",
@@ -158,18 +153,6 @@ class FriendsView(GenericAPIView):
         )
         return paginator.get_paginated_response(serializer.data)
 
-
-@extend_schema(
-    summary="Search Users",
-    description="Search for users by email or username. 유저 검색",
-    tags=["Users"],
-)
-class UserSearchView(ListAPIView):
-    permission_classes = [IsAuthenticated]  # Only allow authenticated users to search
-    serializer_class = UserSearchSerializer
-    queryset = User.objects.all()
-    filter_backends = [filters.SearchFilter]  # Use the search filter backend
-    search_fields = ["email", "username"]  # Fields to search in
 
 
 class FriendRequestView(GenericAPIView):
@@ -295,44 +278,4 @@ class FriendRequestView(GenericAPIView):
         serializer = FriendSerializer(friend)
 
         return Response(serializer.data, status=201)
-
-
-@extend_schema(
-    summary="Search Users who are friendable",
-    description="Search for users by email or username who are not already friends with the current user. Excludes blocked users.",
-    tags=["Users"],
-)
-class SearchFriendableView(ListAPIView):
-
-    permission_classes = [IsAuthenticated]
-    serializer_class = UserSearchSerializer
-    queryset = User.objects.all()
-    filter_backends = [filters.SearchFilter]
-    search_fields = ["email", "username"]
-
-    def get_queryset(self):
-    # Get all friends for the current user (both as user1 and user2)
-        friends_users = Friend.objects.filter(
-            Q(user1=self.request.user) | Q(user2=self.request.user)
-        ).values_list("user1", "user2")  # Get both user1 and user2
-
-        # Flatten the result to a single list and exclude the current user's ID
-        friend_ids = [user for pair in friends_users for user in pair]  # Flattening pairs into a list
-
-        # Exclude users that are already in the friends list
-        queryset = User.objects.exclude(id__in=friend_ids).exclude(id=self.request.user.id)
-        return queryset
-
-    # @extend_schema(
-    #     summary="Search Users",
-    #     description="Search for users by email or username.",
-    #     responses={200: UserSearchSerializer},
-    # )
-    # def get(self, request):
-    #     """
-    #     Return a list of users that can be sent friend requests.
-    #     """
-    #     users = self.filter_queryset(self.get_queryset())
-    #     serializer = UserSearchSerializer(users, many=True)
-    #     return Response(serializer.data)
